backend/old_stuff/ai/requirements.txt
tensorflow==2.16.1
pandas==2.2.3
numpy==1.26.4
pymongo==4.10.1
python-dotenv==1.0.1
